Fix: 복습문제 쿼리문 변경(#31)

+++ back/src/main/java/com/clover/recode/domain/statistics/repository/TodayReviewRepository.java
@@ -13,5 +13,5 @@ public interface TodayReviewRepository {
 
     List<TodayProblemRes> findTodayReviews(LocalDate date, Long statisticsId);
 
-    int countByStatisticsIdAndDate(Long statisticsId, LocalDate date);
+    Long countByStatisticsIdAndDate(Long statisticsId, LocalDate date);
 }
